PP-7588 Start simplifying Cypress stub creation
Start simplifying the process for generating stubs which are pushed to
Mountebank to stub out requests to other services.

This commit demonstrates how this will be done, by moving generating the
stub in the format expected by Mountebank and using our fixtures into
the stub helper files. Previously this was done in the `stubs.js` file,
which made the code quite opaque and confusing.

To allow for this, the `setupStubs` Cypress command will deal with
either:
1. An object with name and opts fields to constuct a stub from the
`stubs.js` file as it does now
2. An already build Mountebank stub which it will not alter and send
directly to Mountebank.

We can then move the stubs over in several commits following this one.
