+++ test/cypress/integration/request-to-go-live/index.cy.test.js
@@ -1,3 +1,4 @@
+
 const userStubs = require('../../stubs/user-stubs')
 const gatewayAccountStubs = require('../../stubs/gateway-account-stubs')
